database/requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
pydantic==2.8.2
scikit-learn==1.5.1
pandas==2.2.2
numpy==2.0.1
joblib==1.4.2
requests==2.32.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1

